spi flash non terminé

+++ inc/stm32spi3.hpp
@@ -19,7 +19,7 @@ public:
 	~STM32SPI3();
 	void init();
 	void sendByte(uint16_t data);
-	uint16_t receiveData();
+	uint8_t receiveData();
 	void assert();
 	void deassert();
 };
